requirements.txt
Flask==1.0.2
Requests==2.22.0
Werkzeug==0.16.0
mysql-connector==2.2.9
mysql-connector-python==8.0.18
configparser==4.0.2




